Plot the result of xAna_monoH_jets.C
Change the variable name in xAna_monoH_jets.C
`hTHINNJetsMatched` to `hTHINnJetsMatched`
to match that in the infile.
Change the value of
`nameFileHead` (to "jets")
`nameFileTail` (to "20200525")
`namesHistogram` (to `std::vector({"hTHINnJetsMatched"})`)
in xAna_monoH_Draw.C (previously known as xAna_monoH_pfMetCorr_Draw.C)
to draw the histograms in the "jets" files.