BranchMounter.h, BranchMounterForUntuplizer.h: Fix and Run
Previously experienced error is caused by
the template speciallization of vector<bool>
in C++ standard libreary, which is
intended to create a reference to the "bits" inside
and the returned address is not the one for a boolean value

This is also why it only happened to
BranchMounterScalarSingle<Bool_t>

The solution is to add a template speciallization for
BranchMounterScalarSingle<Bool_t>
and make it inherits
BranchMounterScalarSingle<Char_t>
`using` most of the members including
Binding vE and funPush
and hack on
the constructors and SetFunPush
to concatenate to those of
BranchMounterScalarSingle<Char_t>

Classes defined in BranchMounterForUntuplizer.h
now run as part of
xAna_monoZ_preselect.C
to further automate the tree->Branch work.
They are currently used on
Jet_* and SoftActivity* branches,
and the use will be extended to
FatJet_*
without having to modify the *.h files.